Un-futurized test/types/range/stonea/countedRangeBy.chpl.
Our current guess is that this test was added early on in the development
of the 'by' operator and was intended to show when it was legal or not.
Today we have a solid understanding of the issues, so this code
has well-known expectations.

Splitting it into two tests, one with a compiler error and the other without.
Although it is unclear whether they are worth keeping to begin with.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19186 3a8e244f-b0f2-452b-bcba-4c88e055c3ca