Fix setter "flattening" bug in function resolution.
During function resolution, for each var function encountered, a
corresponding value function is built and the setter arg is
"flattened" to true for the var function and false for the value
function.  If the call to resolveFns() fails and is tried again, the
value function was being rebuilt from the version of the var function
that was flattened previously, resulting in both the value and var
functions having setter=true.

This commit ensures that the value function is only built once and
both versions are only flattened once.  There is a potential for
compile time performance improvement.

This bug has not been previously encountered probably due to the
limited use of the setter argument.  It only showed up after the
changes made to promote ChapelRT to a proper modules which changed the
module initialization order.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19188 3a8e244f-b0f2-452b-bcba-4c88e055c3ca