Give the communication task's thread its thread private storage.
Certain information about the task running on a thread is kept in that
thread's thread private storage.  This includes the task's ID, and
serial state.  Prior to this, the communication task's thread didn't
have such storage, so a query about, for example, the communication
task's serial state would fail with an internal error. This mod gives
the communication task's thread its private storage, and so enables
getting and setting the data stored there.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19214 3a8e244f-b0f2-452b-bcba-4c88e055c3ca