Never delete the thread private data key.
If a given Chapel program doesn't cause use of a communication task even
though we've made one, we can end up exiting before the comm task really
starts running and we try to make thread private storage for its thread
(see r19214).  If that happens, the comm task's thread may find that we
have already deleted the thread private data key before it even gets a
chance to use it.  An internal error will result.

Fix this by just never deleting the thread private data key.  There is
no real reason beyond tidiness to delete it anyway.

Note that we may want to revisit this if and when we support unilateral
exit in a more structured way, because then we may prefer to make sure
that we either don't make a communication task and thread, or do make
them and always wait for them to quit.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19216 3a8e244f-b0f2-452b-bcba-4c88e055c3ca