Added slicing for DimensionalDist arrays, plus minor changes to Dimensional.
* Changed some fields from 'var' to 'const'.

* Implemented slicing for Dimensional arrays. (No reindexing at the moment.)

** Added the field 'allocDom' to DimensionalArr (the global array descriptor).
   It is similar to 'allocDom' in LocBlockCyclicArr:
For an array that's an alias, 'allocDom' points to the global domain descriptor
of the original array. Otherwise it points to the array's own domain descriptor
(i.e. same as the 'dom' field).

** The global array descriptor of an array alias
copies the local array descriptors from the original array,
instead of creating its own. (No need to create its own.)

** Modified array iterators (serial, follower, dsiSerialWrite) to take
different paths depending on whether the array is an alias or not.

** Added a test (s8.chpl).

* Reworked the argument checking for the DimensionalDist constructor,
to handle more cases gracefully and catch more at compile time.

* Fixed dsiGetIndices() to return a tuple of ranges rather than a domain,
as per the DSI interface requirements.

* Made a warning message more specific.

* Modified the names of the BlockDim (named 'sdist' at present) constructor
arguments to be more self-explanatory (for a primer).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19218 3a8e244f-b0f2-452b-bcba-4c88e055c3ca