Updating qthreads to version 1.7
This updates to the latest qthreads release, version 1.7, which
seems to be working quite well in the examples directory.  I'm
going to check it in for a full test tonight and will remove 1.6
tomorrow if things go well.

There is still a mod we may want to make to the error exit case
for qthreads, but I thought it would be worth running it as-is
for a night first to see if the results differed in any
significant way from what we have been seeing first.  I'll
plan to prepare the mod so that if things seem fine I can check it
in tomorrow.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19221 3a8e244f-b0f2-452b-bcba-4c88e055c3ca