Added compilerAssert(param test: bool, param args...).
This works like assert() but reports an error at compile-time when !test.
Works like compilerError() but checks the condition before reporting an error.

Two quirks here:

* I had to declare a separate proc for each number of arguments
instead of using the declaration shown above.
I (arbitrarily) stopped at 5 arguments; if there are more, the
warning message will contain [...] at the end.
I tried two alternatives which did not work:

  (a) Given a param tuple like 'args' in the above, its detupling
'(...args)' does not yield params, so I was unable to invoke
compilerError on the detupling.

  (b) I could just create a copy of compilerError()
(where detupling magically works), but add an extra argument.
Turns out that the extra argument breaks the magic.

* Secondly, since compilerAssert is yet another level of function call,
the 'errorDepth' argument, if the user passes it, will need to be
incremented by one; the default should be 2. I could have added more
declarations to see whether there is the errorDepth argument (and increment
it by 1) or there isn't (and pass 2 explicitly). But I opted for
fewer declarations. Feel free to implement that if you feel strongly.
(But ideally detupling will be implemented and the number of declarations
will be reduced. Still will probably need 4 decls.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19227 3a8e244f-b0f2-452b-bcba-4c88e055c3ca