Made it an error for a string literal to contain a newline without a backslash.
That is:

  const s = "first line \
             second line";

is a legal string, containing the newline symbol in it. All other
whitespace is preserved, too. Backslash is not part of the string value.

If the backslash is removed, the compiler will report an error.

I added a couple of tests to check this behavior.

Couple of existing tests used to have un-backslashed newline in their strings.
I added backslashes, which should not change the intention of these tests.

There is still a bug (newline-in-string-compiletime-1.future) and
an open question (undefined-backslash-combos-in-strings-1.future).
These are unaffected by this commit.


NOTES

Reporting the "no backslash" error matches the behavior in C (at least gcc).
Preserving the backslashed newline in the string value differs from C. There,
neither the backslash nor the following newline are included in the value.

In the compiler internal representation, a string literal is stored
as a C string (of the type const? char*). The newline is stored
as two characters, '\' and 'n'. Double quotes are escaped with backslashes,
too. This way codegen can generate the C counterpart of a Chapel string value
by printing its internal representation without any conversions.

This does not work for string literals printed at compile time
(e.g. the arguments to compilerError()). See
  newline-in-string-compiletime-1.future

Also, we do not handle gracefully the backslash combinations
that are disallowed in C (e.g. "\w"). See
  undefined-backslash-combos-in-strings-1.future

Coding note: in processTokens.cpp, I added this assignment:
  yytext[0] = '\0';
to avoid useless text in the error message:
  near '"'


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19228 3a8e244f-b0f2-452b-bcba-4c88e055c3ca