Renamed enum 'iterator' to 'iterKind' and arg 'follower' to 'followThis', to match the upcoming PGAS11 paper on leader/follower iterators.
Renamed existing variables 'followThis' to 'myFollowThis'
- for uniformity; feel free to change.

Renamed existing variables 'follower' to 'followThis'
in a few cases when that was not required - again, for uniformity.

I aimed, to some extent, at making grep results as narrow as possible
when grepping for one of:
 'follower'
 'followThis'
 'myFollowThis'

I factored out the uses of the strings (the ones I thought appropriate)
  "iterKind", "leader", "follower", "followThis" in the compiler
into #defines, supposedly so that it's easier to understand
their intentions as referring to the source code/conventions.

TODO: can we fix the 'update-sources' script to take care of this conversion?
(Or at least hint as to where attention is needed?)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19233 3a8e244f-b0f2-452b-bcba-4c88e055c3ca