Adapt qthreads for the recent task table initialization changes.
This makes changes similar to those needed in the FIFO tasking layer
recently to adapt to the recent task table changes.  Specifically, it
allows calls to chpl_task_getId() and chpl_task_{get,set}Serial() in
non-tasks, both very early when the main process is still initializing
the program, and also later in the pthread created as the communication
task.

The content here is courtesy Dylan Stark of the qthreads team (thanks!).
I'm just acting as clerk. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19237 3a8e244f-b0f2-452b-bcba-4c88e055c3ca