Fix optimized on-clauses regressions.
The RAD optimization added a bunch of unnecessary autocopy/autodestroy
calls that were disabling the fast-on optimization.  Manually inserted
temporaries with the "no copy" and "no auto destroy" pragmas.  This
should hopefully become unnecessary once I get back to the reference
counting work.

Added a few new tests to the distribution robustness suite that make
sure this optimization fires for the various distributions.

Unfortunately, it currently does not fire for the cyclic distribution.
The underlying DefaultRectangular arrays are stridable (param) and
that ultimately leads to a call to chpl__diffMod() (during the member
check) which has a halt() statement to catch error case which results
in non-local communication.  It would be nice if we could simply
ignore halt() calls since we know the program is going to terminate
anyways, but we'd have to make sure that we could do a tear down
operation from the handler in gasnet and other comm layers.

For now, the new tests execute a PREDIFF which sets the .good file to
be one that doesn't expect the optimization to fire (same as for
CHPL_COMM=none).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19244 3a8e244f-b0f2-452b-bcba-4c88e055c3ca