Fix for Michael Ferguson's override_overload2 bug.
Reference temporaries were not getting inserted when the method was called
through a virtual method table.  The problem was due to the fact that calls
with dynamic binding use a different form than statically-bound calls.  The 
call expression is changed to a PRIM_VMT_CALL primitive, the method token
argument is replace by the FnSymbol being called, and a call ID (cid) argument
is inserted immediately after that.  The baseExpr field in the call is replaced
by null, so that calls to isResolved() return NULL (false).

The solution was to rewrite the for_formals_actuals macro so that it will
traverse the argument lists of a PRIM_VMT_CALL as well.  The clause that inserts
reference temporaries was also modified to pick up the PRIM_VMT_CALL case as
well as the usual case (indicated by isResolved() returning a non-NULL value).

Since most uses of this macro also test isResolved() (which still depends
only on the value in the baseExpr field), VMT calls are likely unaffected by
this change in most places.  A better long-term solution would be to subclass
the call expression for VMT calls, so special argument-traversal code was not
needed.

DETAILS:

Modified the for_formals_actuals macro to traverse the arguments of a
PRIM_VMT_CALL as well as a "normal" function call.
    compiler/include/alist.h

Expanded the conditional enabling the insertion of reference temporaries to
include PRIM_VMT_CALL primitives along with calls for which isResolved() returns
a non-NULL value. 
    compiler/resolution/functionResolution.cpp


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19247 3a8e244f-b0f2-452b-bcba-4c88e055c3ca