Added scripts that can be used as system-wide prediff scripts for GASNet and bsub.
-----------

prediff-for-amudp-warning:

# This script is for the use with 'start_test' as the system-wide prediff
# when GASNet is used and may produce "Receive buffer full" wwarnings.
# To use (assuming bash):
#
#  CHPL_SYSTEM_PREDIFF=$CHPL_HOME/util/test/prediff-for-amudp-warning \
#    start_test ...

-----------

prediff-for-bsub-launcher -h:

prediff-for-bsub-launcher is for the use with 'start_test' as
the system-wide prediff, when the launcher is bsub.
E.g. (assuming bash):

  CHPL_SYSTEM_PREDIFF=$CHPL_HOME/util/test/prediff-for-bsub-launcher \
    start_test ...

prediff-for-bsub-launcher assumes that the launcher prepends
to the test output three lines that start like this:

  Job XXX is submitted to
  <<Waiting for dispatch
  <<Starting on

If so, they are removed from the test output file.
Otherwise, the output file remains unchanged.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19260 3a8e244f-b0f2-452b-bcba-4c88e055c3ca