Disable bulk transfer optimization on any array aliases for Block distribution.
There is a bug in the iterator for bulk transfer for the Block
distribution that does not properly handle rank changed arrays for
which the last (right-most) dimension is a collapsed dimension.  It
was decided to disable the problem case for the release, since the
iterator is a bit tricky.

Unfortunately, I had to disabled bulk transfer for all aliased arrays
to handle this case (a *gigantic* hammer), because in general it is
not possible to determine the collapsed dimension without visiting at
least one (and possibly all) the locales on which the arrays are
allocated, and that seemed too expensive for the uncommon case.  Since
this is very conservative, you can enable bulk transfer on aliased
arrays (with caveats) by specifying the config const
--disableAliasedBulkTransfer=false.

NOTES: Brad had suggested checking the rank of the _arrAlias in the
array class for the work around, but unfortunately, _arrAlias is
declared as type BaseArr which doesn't include the rank field.  Some
ideas thrown around for future investigation include making _arrAlias
generic or including some basic fields in BaseArr.

TESTING: Added a future test case for this issue that is compiled
with --disableAliasedBulkTransfer=false.  I realized I hadn't checked
in NUMLOCALES file last time, so I added that here also.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19265 3a8e244f-b0f2-452b-bcba-4c88e055c3ca