Made the output of 'diffuse2d-orig-blc --verbose=true' less noisy when the number of cores varies.
The previous .good file worked well with --dataParTasksPerLocale=2.
If the number of cores (and dataParTasksPerLocale) is not 2,
a different number of floating-point operations is performed,
producing slightly different errors under 9e-13.
That made 'start_test' report an error on the test.
This commit fixes that.

The 'eps' var and the check 'err > eps' are back-ported
from diffuse2d-v3-blc.chpl.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19275 3a8e244f-b0f2-452b-bcba-4c88e055c3ca