Change serialStmt1.chpl to be a little less leading towards one of our grey areas.
The scope of the index variable that is introduced in a for/forall
loop is considered a constant, and as a result copies of it are not
made if it is referenced within a begin/cobegin block.  This can lead
to surprising results, and so I've changed this example to avoid such
a situation.  The semantics of this are under inactive discussion.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19303 3a8e244f-b0f2-452b-bcba-4c88e055c3ca