Added chapter on interoperation.
Added external and exported procedure declaration nonterminals.
    spec/Statements.tex

Inserted the Interoperation chapter after Standard_Layouts.
    spec/spec.tex

Moved the definition of config-or-extern here, since it disappeared from
Types.tex.
    spec/Variables.tex

Split class-declaration-statement into external-class-declaration and
simple-class-declaration.  The new nonterminal external-class-declaration is
defined in terms of the other.
At the end of the section explaining extern classes, added a forward referenct
to the Interoperation chapter.
    spec/Classes.tex

Split record-declaration-statement into simple- and
external-record-declaration-statements.
Added a forward reference to the chapter on Interoperation.
    spec/Records.tex

This is the new chapter.
    spec/Interoperation.tex

Split type-alias-declaration-statement into a simple type alias declaration and
an external-type-alias-declaration-statement.
Removed config-or-extern, since it is no longer used here.
At the end of the section on external type alias declarations, added a forward
link to the chapter on Interoperation.
    spec/Types.tex

Rewrote most of the earlier discussion of linkage specifiers.
Added an explanation of why inlining is desirable.
Added a forward reference to the chapter on interoperation.
    spec/Functions.tex

Put the information on creating libraries here, since it is not baked (and also
not complete.  How are they used, e.g.
    doc/release/technotes/README.libraries


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19310 3a8e244f-b0f2-452b-bcba-4c88e055c3ca