Completed my task parallelism follow-ups
Moved memory consistency model into its own chapter and reworded
somewhat to (a) talk about things being more 'undefined' than
'incorrect', (b) describe things in a way that I think intuitively
gets across what we're trying to do (even though we admit to it
being unfinished), (c) added a new example showing the use of a
flag as an indicator that a write had finished, but didn't turn
on its automatic testing yet, due to the code freeze.

Rewrote the atomic section to be a bit shorter and less defined.
Replaced the existing examples with a doubly-linked list example
(that I also didn't turn on as a test due to the code freeze).

Added the new chapter to the spec and organization chapters.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19312 3a8e244f-b0f2-452b-bcba-4c88e055c3ca