Add a basic script that tests Chapel with all the Cray Programming Environments.
This script tests all the PrgEnv modules specified on the command
line.  If none are specified, test gnu, pgi, intel, and cray.  The
script assumes it is run from the examples directory, and that the
directory can be written.

For each PE:
- swap in the PE
- 'module show' PE
- make clean
- make (assumes that chpl is in your path)
- run hello6-taskpar-dist
- run hpcc/stream

There is a small amount of error checking.  The script will report
failure if the module was not found and if the test runs failed.  The
script does not use start_test.

I'm sure this could be better, and I'm sure I'll get annoyed with it
and improve it.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19333 3a8e244f-b0f2-452b-bcba-4c88e055c3ca