Changed my version of HPL to use Dimensional/block-cyclic throughout plus replicated in 1 dimension in schurComplement(). Enabled it as a test.
* Ab is mapped with Dimensional(block-cyclic dim, block-cyclic dim)
  replA,replB are mapped with Dimensional(block-cyclic dim, replicated dim)

* Renamed MatVecSpace -> AbD to support the following.

* The changes for using Dimensional and for schurComplement() are
copied from the current bc.dim.chpl, with only minor mods. In particular:

** Reference the globals instead of passing the following variables
as arguments to LUFactorize() et al.:
  AbD
  Ab
  MatrixSpace (which formal was unused anyway)

The primary motivation is the following compiler glitch: when a
privatized array is passed as an argument and then into a forall
and/or into 'on', its pid is treated as a mutable int and so is
fetched via a remote reference from within the forall/on.  Therefore
without this change the local matrix multiply in schurComplement()
would violate the 'local' requirement.

(I have not rechecked this glitch lately, but I am unaware of any
changes that may have fixed it.)

Once I did that for schurComplement, I did the same everywhere else
for consistency.

** Inlined dgemmNativeInds(), probably as a workaround for the same bug.
The code feels simpler that way, at least to me.

** Removed main() so that the top-level code is now the module initialization.
The reason is because AbD and Ab were declared in main(), but
I need them to be globals (see above).

** Reshape Locales into targetLocales explicitly (based on numlocales).
That's because the Dimensional constructor currently requires
a 2-d array of locales.

** replA, replB are globals, not local to schurComplement().
I think the motivation is to avoid creation overhead.

* panelSolve() is unchanged for now (i.e. essentially the same
as in hpl-blc.chpl I think). The prospective improvements are in comments.

* Added printouts "processing block 1", 9, ...,
to display progress for slow executions.

* Problem size -> 100

* Numlocales -> 4

* Ported the fixes from r19358 (duh).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19362 3a8e244f-b0f2-452b-bcba-4c88e055c3ca