Changed getFailedTestNames to included timed out tests.
At Brad's request, added timed out tests to the mix.

   usage: getFailedTestNames filename [compiler|program|timeout]

It's a bit annoying to pick out the test names because of the varying
number of words in the error message before the test name, so I left
out error messages for missing .good files (since the number wasn't 6
like the others :)

If this is important, we should probably rewrite it to be a bit more
robust.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19363 3a8e244f-b0f2-452b-bcba-4c88e055c3ca