Simple code cleanup for get_int()/get_uint()
This commit changes the compiler's get_int/get_uint calls to pass back
values of type int64_t and uint64_t rather than long/unsigned long as
they were before.  This is smart because we don't want to rely on
integers of poorly-defined width in the compiler, and because the
underlying calls that get_int/uint are built upon pass back the
int64_t/uint64_t calls.

I came across this when using the Mac compiler (which warned about it)
on Michael's I/O branch.  Something he had done there caused gcc on
the Mac to complain even though I wasn't seeing it on trunk.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19373 3a8e244f-b0f2-452b-bcba-4c88e055c3ca