Reverted replication for HPL with Dimensional from r19370 back to r19362.
The changes in r19370 did not increase (decreased, more likely)
the performance when running on 64 nodes on a Cray XE,
while resulting in less elegant code.

Other than replication, r19370 also removed comments with suggestions on
how to replicate the sub-row used in the last computation in panelSolve().
Those comments are still out because they are not essential.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19376 3a8e244f-b0f2-452b-bcba-4c88e055c3ca