Removing an 'inout' intent that generated warnings
Removing an 'inout' intent that was resulting in serialized
domain/array copies and generating a warning.  In email discussions
with Michael, he was thinking/wanting 'ref' when using this (and
arguably he really wants support for varargs type arguments).  Testing
seems to result in strictly better results (30-some failures compared
to 130-some).



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/ferguson@19377 3a8e244f-b0f2-452b-bcba-4c88e055c3ca