Updating test to reflect new I/O approach
Michael's new default I/O routines access fields directly rather than
going through the getter/setter calls which changes the behavior of
this test -- what used to be an illegal get in the writeThis() call
now does not trip the user's test in the getter.

I don't have a strong opinion as to whether or not the default I/O
should use getters/setters, but it seems reasonable that it shouldn't
need to; and this test wasn't really intending to test the I/O
routines per se, more the ability to have a getter/setter that shared
the name of a field.  So I rewrote the test to reflect the current
behavior while still tripping the error condition.

Note that the way to cause the I/O routines to use the getters/setters
would probably be to have the fielderator primitives use a
getter/setter function call expression rather than calling the
PRIM_GET_MEMBER routine as it currently does.




git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/ferguson@19378 3a8e244f-b0f2-452b-bcba-4c88e055c3ca