Kill the internal module circular-dependency bug.
There was a mysterious codependency between ChapelStandard and ChapelBase which
has caused repeated difficulties for people modifying the internal modules --
particularly ChapelLocale, ChapelBase and DefaultRectangular.

I traced the problem to line 170 of filesToAst.cpp, which is where "use
ChapelStandard" is implicitly added to every module except a select few.  The
set of exempted modules should have included ChapelBase but didn't.  I merely
added the missing module to the list.

DETAILS:

Added baseModule to the list of modules which *do not* depend on ChapelStandard.
    compiler/passes/filesToAST.cpp

Renamed the test case Writer.chpl as UserWriter.chpl to work around a bug.
The bug is that when there is an argument declaration whose declared type is a
class type whose definition has not yet been provided AND there exists a module
with a matching name, the ModuleSymbol is used instead of the TypeSymbol.  

This is obviously wrong, but a bit remote from the current fix.  So I added the
workaround alongside a test case to track the bug.
    spec/Input_and_Output.tex

Added a future to track the name conflict error which appeared in the Writer
test case.
    test/modules/hilde/Writer.chpl
    test/modules/hilde/Writer.good
    test/modules/hilde/Writer.future


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19380 3a8e244f-b0f2-452b-bcba-4c88e055c3ca