Removing hack to get cce/8.0
You may recall that a few weeks back I hacked up these test scripts
in order to manually swap in a newer version of cce to avoid a bug
in version 7.x.  Now that cce/8.0 is the default, I'm removing that
hack so that we're always testing the latest and greatest.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19384 3a8e244f-b0f2-452b-bcba-4c88e055c3ca