Use the C++ compiler for "make depend" on the tcmalloc interface source file.
The interface between the runtime's tcmalloc memory layer and the tcmalloc
package is runtime/src/mem/tcmalloc/tcmalloc-interface.cc, which is a C++
file.  We need to use the C++ compiler when we're making the dependency
information for this file.

(Note that as things stand right now, gcc/g++ is the only tested gen compiler
for the tcmalloc third party package, and thus its interface module in the
runtime.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19397 3a8e244f-b0f2-452b-bcba-4c88e055c3ca