Added 4 variations of a bug whereby the compiler does something wrong when a dynamically dispatched method has a param argument (or even doesn't).
From .futures:

* FYI, seems like the compiler creates a method

  BaseArr.getAllocationDomainForDefaultRectangularArray()

as a statically-dispatched wrapper for dynamic dispatch, from which

  DefaultRectangular.getAllocationDomainForDefaultRectangularArray()

is invoked when needed.

* Inside one of the compiler-generated clones (probably the one mentioned
right above) of getAllocationDomainForDefaultRectangularArray, I observed
that another getAllocationDomainForDefaultRectangularArray is passed the
'rank' param. However, by that time, 'rank' has already been eliminated
when specializing the method, and so is no longer available.

So maybe the bug is merely that instead of passing 'rank' we should
pass its corresponding value?

* When compiling with --explain-call xxx, the following line is shown twice:

  generic-method-with-param-arg-3.chpl:24: note: call: xxx(BaseArr)

Just like with generic-method-with-param-arg-2.chpl, the first time
it finds visible functions and resolves it twice, the second time it doesn't.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19399 3a8e244f-b0f2-452b-bcba-4c88e055c3ca