Use a pragma to control implicit inclusion of ChapelStandard.
The decision of whether to use ChapelStandard implicitly was being made on line
170 of filesToAst.cpp.  Internally-generated modules and modules loaded by
parseInternalModule() were stored in global module symbols, and matching on
these symbols would allow the normal inclusion of ChapelStandard to be skipped.

With this change, implicit use of ChapelStandard can be inhibited by attaching
the "no use ChapelStandard" pragma to the module.  This change was propagated to
all existing internal modules.  Additional use statements were added as needed
to resolve compile-time dependencies.  (Due to an existing bug, type names which
are resolved at compile time (during folding) and are unknown at the point of
use are erroneously resolved as module names, which leads to failure in the
normalize pass.)

ASIDE:

This checkin is not expected to alter the following behavior, but it would be
nice to understand nonetheless.  Function symbols which are referenced in user
modules must be defined by an explicit use in a scope reachable from the
reference.  In contrast, function symbols defined in internal modules can
seemingly be used anywhere.

Consider:
    ----------------------------------------------------------------
==> x1.chpl <==
use x2;
use x3;
==> x2.chpl <==
proc test() { writeln("I am in test"); }
==> x3.chpl <==
test();

% chpl x1.chpl
./x3.chpl:1: error: unresolved call 'vass()' [functionResolution.cpp:1598]
    ----------------------------------------------------------------
and now compare:
    ----------------------------------------------------------------
==> ChapelBase.chpl <==
...
proc _isIntegralType(type t) param return
  _isSignedType(t) || _isUnsignedType(t);
...
==> ChapelUtil.chpl <==
...
proc safeSub(a: ?t, b: t) {
  if !_isIntegralType(t) then
...
    ----------------------------------------------------------------
Why is _isIntegralType legal in ChapelUtil.chpl but test() is not legal in x3.chpl?



DETAILS:

Add the "no use ChapelStandard" pragma.
Internally called FLAG_NO_DEFAULT_USE.
    compiler/include/flags_list.h

Attach the new flag to internally generated modules rootModule and theProgram.
    compiler/AST/type.cpp

On line 170, we now filter modules that don't want to include ChapelStandard
using the new FLAG_NO_DEFAULT_USE flag.
In parserInternalModule(), we now have to look up the module by name.
This is because ParseMod() only returns a module symbol if the entire file is
one (unwrapped) module.  We are now wrapping internal modules so ParseMod()
returns NULL.  It might be preferable to push this down into ParseMod().  If the
file contains a single module, then whether wrapped or not we can return that
one module.  Then, the linear search of module names in parseInternalModule()
can be eliminated.
    compiler/passes/filesToAST.cpp

Add the "no use ChapelStandard" pragma to module headers.
Add uses for other specific dependencies as needed.
    modules/internal/ChapelUtil.chpl
    modules/internal/ChapelArray.chpl
    modules/internal/ChapelDistribution.chpl
    modules/internal/DefaultOpaque.chpl
    modules/internal/ChapelTasksInternal.chpl
    modules/internal/DefaultSparse.chpl
    modules/internal/ChapelRangeBase.chpl
    modules/internal/ChapelThreadsInternal.chpl
    modules/internal/ChapelTuple.chpl
    modules/internal/ChapelThreads.chpl
    modules/internal/ChapelRange.chpl
    modules/internal/DefaultRectangular.chpl
    modules/internal/ChapelLocale.chpl
    modules/internal/ChapelStandard.chpl
    modules/internal/DefaultAssociative.chpl
    modules/internal/ChapelNumLocales.chpl
    modules/internal/ChapelReduce.chpl
    modules/internal/ChapelIO.chpl
    modules/internal/ChapelBase.chpl
    modules/internal/ChapelTaskTable.chpl


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19400 3a8e244f-b0f2-452b-bcba-4c88e055c3ca