Standalone module initializers.
Reworked the module initialization code so that each module calls only the
initializers for the modules on which it depends directly.  The initialization
guards ensure that each module initializer is called only once.  When separate
compilation is more fully implemented, the ability to call a module initializer
in a standalone fashion will be necessary.

An important supporting change involves resolving module initialization
functions in dependency order before chpl_main is resolved, and then adding the
module initializer calls to module initialization routines in the addInitGuards
pass (just before code generation).

The insertion of the module initializer calls (corresponding to each "use"
statement) was deferred until late in the compilation, because otherwise the
localizeGlobals phase inserts extra initialization code in front of them.  These
variable initializations can depend on values established in used modules, so
that puts them in the wrong order.  Adding the initializer calls late in the
compilation makes sure the module initializers are called first.

Resolving module initializers in dependency order ensures that all of the
modules used by the program will be initialized in the order required by the
spec.  This used to be done by traversing the dependency list and dropping the
module initializer calls into chpl__init_chpl__Program() and then making sure
that chpl_main called this.  Resolution of chpl_main would resolve dependencies
in the right order.

A side-effect of deferring insertion of the module initializer calls until after
resolution is that the depth of the call stack changes in diagnostic messages.
Before, chpl_main was the root of the callstack.  Now, each of the module
initialization routines plus chpl_main are roots.  (Actually,
standardModule->initFn and every exported function also used to be roots.)

DETAILS
-------

The recursive function resolveUses() now resolves all of the module
initialization functions (in dependency order).
Now use vector head() and tail() routines.
4013: It was assumed that if you popped back to level 0 in the call stack, then
no more resolution could be done.  That was true when chpl_main was root of the
call tree, because it contained no code.  But when each module initializer is
resolved separately, we can resolve symbols in level 0 as well.  This would have
created in a bug in exported routines (since they are also roots), but none of
our regression tests were complex enough code to expose the problem.
4653: We no longer resolve starting from chpl__init_chpl__Program() but from
each of the module initialization functions (in dependency order) and then
finally the whole program.
    compiler/resolution/functionResolution.cpp

In the runtime, we now call chpl__init_chplStandard instead of
chpl__init_chpl__Program.  The latter routine has become moot.
    runtime/include/chplcgfns.h
    runtime/src/main.c

A slight clarafication of module initialization order.
    spec/Modules.tex

mainModule was made global, so it can be used as the root of module dependency
resolution.
    compiler/include/symbol.h
    compiler/AST/symbol.cpp

Renamed addInitGuards() as addInitCalls().
    compiler/include/passes.h
    compiler/main/passlist.h

Added head() and tail() functions.  The idiom vector.v[vector.n-1] was used a
lot, so it seemed right to factor that out as tail () to make it more
self-descriptive.  Then, head() was added for symmetry.
    compiler/include/vec.h

Regularization:  No need to keep from creating a module initialization block for
theProgram.  The block is empty, so behavior is unchanged.
    compiler/AST/build.cpp

Renamed the top-level function as addInitCalls().  This now calls
addModuleInitBlocks() and then addInitGuards().  These put module initialization
calls and guards in the right place in each module initializer function.
Now, we iterate over modules and select their initialization functions rather than
iterating over all functions and selecting just the module initializer
functions.  One step closer to retiring FLAG_MODULE_INIT.
    compiler/passes/addInitGuards.cpp

Adding module initialization calls was moved from here to addInitGuards.cpp.
    compiler/passes/buildDefaultFunctions.cpp

It would be nice to be able to mark every module initialization routine as
exported and retire FLAG_MODULE_INIT.  But currently it creates too many name
conflicts.  (Module names do not have to be globally unique, so we need a
mangling scheme for exported names.)
    compiler/AST/astutil.cpp

Instantiation order changed.
    test/classes/hilde/explainInstantiation.good

Call stack depth changed.
    test/execflags/vass/print-callstack-on-error-1.good
    test/execflags/vass/print-callstack-on-error-2.good
    test/execflags/vass/print-callstack-on-warning-1.good
    test/execflags/vass/print-callstack-on-warning-2.good
    test/functions/diten/compilerWarningDepth.chpl
    test/functions/diten/compilerWarningDepth.good

Mystery improvement: we now get a real line number.
    test/functions/bradc/paramFnNonParamArgs/paramExample.1-0.good


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19401 3a8e244f-b0f2-452b-bcba-4c88e055c3ca